Switch to calico 1.2.0 and use IP_AUTODETECTION_METHOD
When running on virtualbox, one of the interfaces
is a NAT interface, and if the calico IP autodetect
code picks that interface, things don't work.  So,
update to calico 1.2.0 so that we can use the
IP_AUTODETECTION_METHOD feature to tell calico
how to find the correct IP address.  The IP
addresses are in the same subnet as is used by
the Vagrant configuration.